# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoneyPrinterTurbo",
    "description": "<a href='https://github.com/harry0703/MoneyPrinterTurbo'>https://github.com/harry0703/MoneyPrinterTurbo</a>",
    "version": "1.3.7"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "List projects",
        "operationId": "get_projects_api_v1_projects_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "V1"
        ],
        "summary": "Create a project",
        "operationId": "create_project_api_v1_projects_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episodes": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "List episodes",
        "operationId": "get_episodes_api_v1_episodes_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episodes/{episode_id}": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "Get one episode",
        "operationId": "get_episode_api_v1_episodes__episode_id__get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Episode / task ID",
              "title": "Episode Id"
            },
            "description": "Episode / task ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "V1"
        ],
        "summary": "Move an episode to another project",
        "operationId": "update_episode_api_v1_episodes__episode_id__patch",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Episode / task ID",
              "title": "Episode Id"
            },
            "description": "Episode / task ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episodes/{episode_id}/scenes": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "List an episode's scenes",
        "operationId": "get_scenes_api_v1_episodes__episode_id__scenes_get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Episode / task ID",
              "title": "Episode Id"
            },
            "description": "Episode / task ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episodes/{episode_id}/assets": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "List an episode's source materials",
        "operationId": "get_assets_api_v1_episodes__episode_id__assets_get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Episode / task ID",
              "title": "Episode Id"
            },
            "description": "Episode / task ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episodes/{episode_id}/renders": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "List an episode's finished videos",
        "operationId": "get_renders_api_v1_episodes__episode_id__renders_get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Episode / task ID",
              "title": "Episode Id"
            },
            "description": "Episode / task ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/logs": {
      "get": {
        "tags": [
          "V1"
        ],
        "summary": "Read a task's captured logs",
        "description": "Logs are an in-memory ring buffer in the process that ran the task\n(`app/services/task_logs.py`), capped at the 20 most recent tasks. A task\nstarted in a different process -- the Streamlit page, say -- has no logs\nhere, and neither does one whose buffer has since been evicted. Both cases\nreturn an empty list rather than a 404.",
        "operationId": "get_task_logs_api_v1_tasks__task_id__logs_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Task ID",
              "title": "Task Id"
            },
            "description": "Task ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssetListData": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetData"
            },
            "type": "array",
            "title": "Assets"
          }
        },
        "type": "object",
        "required": [
          "assets"
        ],
        "title": "AssetListData"
      },
      "AssetListResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/AssetListData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "AssetListResponse"
      },
...
      "EpisodeDetailData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "topic": {
            "type": "string",
            "title": "Topic",
            "default": ""
          },
          "script": {
            "type": "string",
            "title": "Script",
            "default": ""
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "run_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Run Data",
            "default": {}
          },
          "state": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "project_id"
        ],
        "title": "EpisodeDetailData"
      },
      "EpisodeDetailResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/EpisodeDetailData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "EpisodeDetailResponse"
      },
      "EpisodeListData": {
        "properties": {
          "episodes": {
            "items": {
              "$ref": "#/components/schemas/EpisodeSummary"
            },
            "type": "array",
            "title": "Episodes"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "episodes",
          "total",
          "limit",
          "offset"
        ],
        "title": "EpisodeListData"
      },
      "EpisodeListResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/EpisodeListData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "EpisodeListResponse"
      },
...
      "EpisodeUpdateRequest": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "EpisodeUpdateRequest"
      },
      "EpisodeUpdateResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "type": "null",
            "title": "Data"
          }
        },
        "type": "object",
        "title": "EpisodeUpdateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreateRequest"
      },
      "ProjectCreateResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ProjectData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ProjectCreateResponse"
      },
      "ProjectData": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "episodes": {
            "type": "integer",
            "title": "Episodes",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectData"
      },
      "ProjectListData": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectData"
            },
            "type": "array",
            "title": "Projects"
          }
        },
        "type": "object",
        "required": [
          "projects"
        ],
        "title": "ProjectListData"
      },
      "ProjectListResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ProjectListData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ProjectListResponse"
      },
...
      "RenderListData": {
        "properties": {
          "renders": {
            "items": {
              "$ref": "#/components/schemas/RenderData"
            },
            "type": "array",
            "title": "Renders"
          }
        },
        "type": "object",
        "required": [
          "renders"
        ],
        "title": "RenderListData"
      },
      "RenderListResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/RenderListData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "RenderListResponse"
      },
...
      "SceneListData": {
        "properties": {
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/SceneData"
            },
            "type": "array",
            "title": "Scenes"
          }
        },
        "type": "object",
        "required": [
          "scenes"
        ],
        "title": "SceneListData"
      },
      "SceneListResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/SceneListData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SceneListResponse"
      },
...
      "TaskLogData": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "logs"
        ],
        "title": "TaskLogData"
      },
      "TaskLogResponse": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 200
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/TaskLogData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "TaskLogResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}